requirements.txt
sentence-transformers
faiss-cpu
scikit-learn
nltk
pymupdf
streamlit
pandas
numpy
plotly
beautifulsoup4
requests
google-search-results
seaborn
matplotlib
reportlab
python-dotenv
serpapi
anthropic